requirements.txt
pymongo
python-dotenv
faker
requests